factor out `GithubRenderer`, avoid `expect_ruff` calls
this prepares to move the code to `ruff_db` by using a `&dyn FileResolver`
instead of the `EmitterContext` directly and by handling the cases where a span
or range is missing. I think this also fixes a latent bug where we were only
falling back on a default `LineColumn` for notebooks for the start location.
`end_location` was used directly without respecting the `is_notebook` check.

note that the one remaining `unwrap` call will disappear after moving the code
to `ruff_db`. the `UnifiedFile` method we need is still private