[ruff] Move `Locator` out of `ruff_linter` and to `ruff_source_file`
The motivation here is to eventually more the "insert import
statements" code out of `ruff_linter` and into its own crate.
But it uses a `Locator` which itself encapsulates lazy construction
of a `LineIndex`. So it seems wise to punt it out.

Since `ruff_source_file` owns `LineIndex`, this is arguably a
better home for it anyway.