Stabilize `unused-unpacked-variable` (`RUF059`)
The tests looked good. For the docs, I added a `## See also` section pointing to
the closely-related F841 (unused-variable) and the corresponding section to F841
pointing back to RUF059. It seems like you'd probably want both of these active
or at least to know about the other when reading the docs.