Stabilize `pytest-raises-ambiguous-pattern` (`RUF043`)
This one has been a bit contentious in the past. It usually uncovers ~700
ecosystem hits. See:

- https://github.com/astral-sh/ruff/pull/16657
- https://github.com/astral-sh/ruff/issues/16690

But I think there's consensus that it's okay to merge as-is. We'd love an
autofix since it's so common, but we can't reliably tell what a user meant. The
pattern is ambiguous after all :laughing:

This is the first rule that actually needed its test case relocated, but the
docs looked good.