Simplify `Program` configuration for GitHub output (#20420)
## Summary

This is a follow-up to https://github.com/astral-sh/ruff/pull/20358 to
avoid adding a new `DisplayDiagnosticConfig` option (with a surprising
default value) just to print the program name for GitHub diagnostics.

Initially I expected this to be a larger refactor including making the
various `Renderer::render` methods take a `std::io::Write` instead of a
`std::fmt::Formatter` to make them easier to call directly in Ruff.
However, this was a bit more painful than expected because we use the
`Display` implementation of `DisplayDiagnostics` for more than writing
directly to stdout. Implementing `Display` on the individual renderers
seemed like a smaller step in this direction.

At this point it might make sense just to merge this into
https://github.com/astral-sh/ruff/pull/20358. Alternatively, I could
apply the same transformation to the other affected renderers. For
example, the JSON and JSON-lines formats only use their `config` fields
for the `preview` setting. Or we could update all of the renderers and
call them directly in Ruff instead of assembling a config at all. I just
wanted to make sure this approach looked reasonable before applying it
to everything.

## Test Plan

Existing tests